tech: Add more tests after upgrading pactffi to v0.4
Also, break up the matcher code into smaller chunks
