+++ PactSwift.xcfilelist
@@ -1,64 +1,30 @@
 $(SRCROOT)/Sources/Extensions/Bundle+PactSwift.swift
-$(SRCROOT)/Sources/Extensions/Date+PactSwift.swift
-$(SRCROOT)/Sources/Extensions/UUID+PactSwift.swift
-$(SRCROOT)/Sources/Extensions/String+PactSwift.swift
-$(SRCROOT)/Sources/Extensions/Dictionary+PactSwift.swift
-$(SRCROOT)/Sources/Extensions/Sequence+PactSwift.swift
 $(SRCROOT)/Sources/Toolbox/Logger.swift
 $(SRCROOT)/Sources/Toolbox/PactFileManager.swift
 $(SRCROOT)/Sources/Model/Constants.swift
-$(SRCROOT)/Sources/Model/Metadata.swift
 $(SRCROOT)/Sources/Model/ProviderVerifier+Error.swift
-$(SRCROOT)/Sources/Model/TransferProtocol.swift
-$(SRCROOT)/Sources/Model/PactPathParameter.swift
+$(SRCROOT)/Sources/Model/Interaction+Extensions.swift
 $(SRCROOT)/Sources/Model/PactInteractionNode.swift
+$(SRCROOT)/Sources/Model/HeaderBuilder+Extensions.swift
 $(SRCROOT)/Sources/Model/ProviderVerifier+Provider.swift
-$(SRCROOT)/Sources/Model/PactHTTPMethod.swift
 $(SRCROOT)/Sources/Model/PactSwiftVersion.swift
-$(SRCROOT)/Sources/Model/Pact.swift
 $(SRCROOT)/Sources/Model/ProviderVerifier+Options.swift
-$(SRCROOT)/Sources/Model/ProviderState.swift
-$(SRCROOT)/Sources/Model/Response.swift
-$(SRCROOT)/Sources/Model/PactInteractionElement.swift
-$(SRCROOT)/Sources/Model/Interaction.swift
-$(SRCROOT)/Sources/Model/Request.swift
-$(SRCROOT)/Sources/Model/EncodingError.swift
-$(SRCROOT)/Sources/Model/ExampleGeneratorExpressible.swift
+$(SRCROOT)/Sources/Model/BodyBuilder+Extensions.swift
+$(SRCROOT)/Sources/Model/QueryBuilder+Extensions.swift
 $(SRCROOT)/Sources/Model/PactBroker.swift
 $(SRCROOT)/Sources/Model/ErrorReportable.swift
 $(SRCROOT)/Sources/Model/WIPPacts.swift
-$(SRCROOT)/Sources/Model/Pacticipant.swift
-$(SRCROOT)/Sources/Model/MatchingRuleExpressible.swift
 $(SRCROOT)/Sources/Model/ErrorReporter.swift
-$(SRCROOT)/Sources/Model/Toolbox.swift
-$(SRCROOT)/Sources/Model/AnyEncodable.swift
 $(SRCROOT)/Sources/Model/VersionSelector.swift
-$(SRCROOT)/Sources/PactBuilder.swift
-$(SRCROOT)/Sources/Matchers/FromProviderState.swift
-$(SRCROOT)/Sources/Matchers/OneOf.swift
-$(SRCROOT)/Sources/Matchers/SomethingLike.swift
-$(SRCROOT)/Sources/Matchers/EqualTo.swift
-$(SRCROOT)/Sources/Matchers/EachLike.swift
-$(SRCROOT)/Sources/Matchers/EachKeyLike.swift
-$(SRCROOT)/Sources/Matchers/IntegerLike.swift
+$(SRCROOT)/Sources/Matchers/Matcher+V4.swift
+$(SRCROOT)/Sources/Matchers/GenericMatcher.swift
+$(SRCROOT)/Sources/Matchers/UUIDFormat.swift
+$(SRCROOT)/Sources/Matchers/Matcher+V2.swift
+$(SRCROOT)/Sources/Matchers/Matcher+Generator.swift
+$(SRCROOT)/Sources/Matchers/GeneratorType.swift
+$(SRCROOT)/Sources/Matchers/Matcher+V1.swift
+$(SRCROOT)/Sources/Matchers/Matcher+V3.swift
 $(SRCROOT)/Sources/Matchers/Matcher.swift
-$(SRCROOT)/Sources/Matchers/RegexLike.swift
-$(SRCROOT)/Sources/Matchers/IncludesLike.swift
-$(SRCROOT)/Sources/Matchers/DecimalLike.swift
-$(SRCROOT)/Sources/Matchers/MatchNull.swift
-$(SRCROOT)/Sources/PFMockService.swift
+$(SRCROOT)/Sources/Matchers/HTTPStatus.swift
+$(SRCROOT)/Sources/Matchers/AnyMatcher.swift
 $(SRCROOT)/Sources/ProviderVerifier.swift
-$(SRCROOT)/Sources/ExampleGenerators/RandomBool.swift
-$(SRCROOT)/Sources/ExampleGenerators/RandomString.swift
-$(SRCROOT)/Sources/ExampleGenerators/RandomDateTime.swift
-$(SRCROOT)/Sources/ExampleGenerators/RandomDecimal.swift
-$(SRCROOT)/Sources/ExampleGenerators/RandomHexadecimal.swift
-$(SRCROOT)/Sources/ExampleGenerators/RandomInt.swift
-$(SRCROOT)/Sources/ExampleGenerators/DateTimeExpression.swift
-$(SRCROOT)/Sources/ExampleGenerators/RandomTime.swift
-$(SRCROOT)/Sources/ExampleGenerators/RandomUUID.swift
-$(SRCROOT)/Sources/ExampleGenerators/ExampleGenerator.swift
-$(SRCROOT)/Sources/ExampleGenerators/RandomDate.swift
-$(SRCROOT)/Sources/ExampleGenerators/ProviderStateGenerator.swift
-$(SRCROOT)/Sources/ExampleGenerators/DateTime.swift
-$(SRCROOT)/Sources/MockService.swift
