+++ Sources/Matchers/Matcher+V1.swift
@@ -1,5 +1,5 @@
 //
-//  Created by Oliver Jones on 9/1/2023.
+//  Created by Oliver Jones on 12/1/2023.
 //  Copyright © 2023 Oliver Jones. All rights reserved.
 //
 //  Permission to use, copy, modify, and/or distribute this software for any
@@ -16,23 +16,18 @@
 //
 
 import Foundation
-import PactSwiftMockServer
 
-public extension HeaderBuilder {
-	/// Set the `Content-Type` header.
-	@discardableResult
-	func contentType(_ contentType: String) throws -> Self {
-		try header("Content-Type", value: contentType)
-	}
+// MARK: - Pact Specification v1 matchers
 
-	@discardableResult
-	func header(_ name: String, value: String) throws -> Self {
-	   try header(name, values: [value])
-	}
+public extension Matcher {
 
-	@discardableResult
-	func header(_ name: String, matching: AnyMatcher) throws -> Self {
-	   let valueString = try String(data: JSONEncoder().encode(matching), encoding: .utf8)!
-	   return try header(name, values: [valueString])
+	/// A matcher that checks that the values are equal.
+	///
+	/// - Note: Requires `Pact.Specification.v1`.
+	/// - Parameters:
+	///   - value: The value to match with.
+	///
+	static func equals<T: Encodable>(_ value: T) -> AnyMatcher {
+		GenericMatcher(type: "equality", value: value).asAny()
 	}
 }
